pyproject.toml
[tool.poetry]
name = "jam-tayang-pro-bot"
version = "1.0.0"
description = "Professional Telegram Bot for Social Media Services"
author = "Kantong Aplikasi 2025"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "20.7"
asyncio = "^3.4.3"
aiohttp = "3.9.1"
sqlalchemy = "2.0.23"
alembic = "1.13.1"
redis = "5.0.1"
requests = "2.31.0"
beautifulsoup4 = "4.12.2"
selenium = "4.15.2"
fake-useragent = "1.4.0"
python-dotenv = "1.0.0"
schedule = "1.2.0"
psutil = "5.9.6"
cryptography = "41.0.8"
pillow = "10.1.0"
qrcode = "7.4.2"
matplotlib = "3.8.2"
pandas = "2.1.4"
numpy = "1.26.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
